+++ src/lib.rs
@@ -308,3 +308,4 @@ pub mod vval;
 pub mod parser;
 pub mod compiler;
 pub mod prelude;
+pub mod threads;